Get started on some items

+++ src/include/item.h
@@ -3,8 +3,9 @@
 #ifndef _item_h_INCLUDE
 #define _item_h_INCLUDE
 
-// Items are divided into 4 blocks of 64 IDs each, 
+// Items are divided into 4 blocks of 64 IDs each.
 // Not sure if I'm gonna keep the 'key' group, but for now I'll put it there.
+// If I remove the 'key' group, I might just expand other groups.
 
 #define ITEM_INDEX_BASE_KEY 0x00
 #define ITEM_INDEX_BASE_CONSUMABLE 0x40
@@ -19,14 +20,14 @@
 
 
 typedef struct {
-    const char* name;
+    const char* name; // Max 20 ch
 } KeyItem;
 
 typedef struct {
-    const char* name;
+    const char* name; // Max 20 ch
 
-    // How much the item heals, or when negative, how much it hurts.
-    int16_t healing_amount;
+    // How much the item heals or, when negative, how much it hurts.
+    int8_t healing_amount;
 
     // How many turns the timed effects of this item lasts.
     // When not actively in an encounter, effects are ignored.
@@ -41,8 +42,8 @@ typedef struct {
 } ConsumableItem;
 
 typedef struct {
-    const char* name;
-    const char* description;
+    const char* name; // Max 20 ch
+    const char* description; // Max 40 ch
 
     // Requirements to use the weapon. If either isn't met, then damage is ignored.
 
@@ -60,8 +61,8 @@ typedef struct {
 } WeaponItem;
 
 typedef struct {
-    const char* name;
-    const char* description;
+    const char* name; // Max 20 ch
+    const char* description; // Max 40 ch
 
     // Weight of the armor. Speed of the wearer is reduced by `max(0, weight - strength)`.
     uint8_t weight;
